% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassData.R
\name{biomassData}
\alias{biomassData}
\title{Extracts and exports fishery independent biomass and abundance data}
\usage{
biomassData(path, s.strat = 440, e.strat = 495, s.year, e.year,
  vessel.correction = TRUE)
}
\arguments{
\item{path}{Filepath indicating where to create folders to store the extracted
data.}

\item{s.strat}{Stratum for which to begin data extraction. Default is
\code{s.strat = 440}.}

\item{e.strat}{Stratum for which to end data extraction. Default is
\code{e.strat = 495}.}

\item{s.year}{Year for which to begin data extraction.}

\item{e.year}{Year for which to end data extraction.}

\item{vessel.correction}{Logical value indicating whether to apply vessel
correction to \code{BIOMASS} and \code{ABUNDANCE}. Default is
\code{vessel.correction} = TRUE.}
}
\value{
Creates directories to store extracted data.

 **Not length-based** path/data/aggregate/ stores an RData file each year
 called year.RData (object name \code{dat}). \code{dat} has 9 columns:
 \code{MISSION}, \code{SETNO}, \code{SPECIES}, \code{YEAR}, \code{STRAT},
 \code{BIOMASS}, \code{ABUNDANCE}, \code{QBIOMASS} and \code{QABUNDANCE}.

 **Length-based** path/data/length/ stores an RData file for each year called
 num_biom_at_length_year.RData (object name \code{out}). \code{out} has 10
 columns: \code{MISSION}, \code{SETNO}, \code{SPECIES}, \code{YEAR},
 \code{STRAT}, \code{BIOMASS}, \code{ABUNDANCE}, \code{QBIOMASS} and
 \code{QABUNDANCE}, and \code{FLEN}, where \code{FLEN} is length in 1 cm
 increments.

 Biomass units are: XXX Abundance units are: XXXX
}
\description{
Extracts and exports annual length-based and not length-based
 (i.e., aggregated over length) biomass and abundance data for the Scotian
 Shelf. \code{QABUNDANCE} and \code{QBIOMASS} are q-adjusted at the set level
 (i.e., before stratification).

 Herring biomass/abundance is treated separately from the other species.
}
\details{
User must define \code{channel = odbcConnect("ptran", uid = ###, pwd
 = ###)} in the global environment. This channel must have access to the XXXX
 databases.

 If \code{BIOMASS} or \code{ABUNDANCE} data is missing when the other is
 present, the missing field is filled in based on the relationship
 \eqn{Weight_{Avg} = \code{BIOMASS}/\code{ABUNDANCE}}. \eqn{Weight_{Avg}} is
 from a sql call to "mean_wts_for_fill_in." If \eqn{Weight_{Avg}} if not
 available for a species, it is assumed that \eqn{Weight_{Avg} = 0.5}.
}
\references{
Modified code from AC's ExtractIndicators/R/biomassData.R
}
