% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineAreas.R
\name{defineAreas}
\alias{defineAreas}
\title{Assigns area ID based on strata}
\usage{
defineAreas(dat, area)
}
\arguments{
\item{dat}{A dataframe that includes the column \code{STRAT}, which has
entries indicating the strat on the Scotian Shelf (e.g., values from 440 to
495).}

\item{area}{Character string indicating the spatial scale for which to assign
areas IDs. Options are: "strat", "nafo", "esswss",  "shelf", "indiseas",
"pollockwest", "pollockeast", "unit2redfish", "unit3redfish", "silverhake".}
}
\value{
The function returns dat with an extra column: \code{ID}
}
\description{
This function adds an area \code{ID} column to a dataframe that
 already has a column \code{STRAT}.
}
\references{
Modified code from AC's ExtractIndicators/R/defineGroups.R.
}
