% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLAND.R
\name{extractLAND}
\alias{extractLAND}
\title{Extracts commercial landings data from historic and current databases}
\usage{
extractLAND(path)
}
\arguments{
\item{path}{Filepath indicating where to create the folder to store the
extracted data.}
}
\value{
This function creates directory path/data/landings and stores file
 landings.Rdata (object name is \code{landings}). \code{landings} has 14
 columns: \code{SPECIES}, \code{ALLNAMES}, \code{YEAR}, \code{NAFO_UNIT},
 \code{CATCH}, and 9 landings groups.
}
\description{
Extracts commercial landings data from NAFO (1968 - 1985), ZIF
 (1986 - 2002) and MARFIS (2003 - present) databases.
}
\references{
Modified code from AC's ExtractIndicators/R/getLandings.R
}
