% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRV.R
\name{extractRV}
\alias{extractRV}
\title{Calls functions to extract and stratify fishery independent data}
\usage{
extractRV(path, s.year, e.year, lengthbased, qadjusted,
  areas = c("shelf", "esswss", "nafo", "strat"))
}
\arguments{
\item{path}{Filepath indicating where to create folders to store the extracted
data.}

\item{s.year}{Year for which to begin data extraction.}

\item{e.year}{Year for which to end data extraction.}

\item{lengthbased}{Logical parameter indicating whether to return stratified
length-based data (\code{lengthBased = TRUE}) or not length-based data
(\code{lengthBased = FALSE}).}

\item{qadjusted}{Logical parameter indicating whether to return q-adjusted
biomass and abundance data (\code{qadjusted = TRUE}) or not q-adjusted
biomass and abundance data (\code{qadjusted = FALSE}). q-adjustments are
applied at the set level.}

\item{areas}{Areas of interest. A separate dataframe will be extracted for
each area. Default is \code{areas = c("shelf", "esswss", "nafo", "strat")}.}
}
\value{
Returns data extracted and stratified by \code{biomassData()} and
 \code{stratifyBiomass()}.
 
 From \code{biomassData()}:
 
 path/data/aggregate stores BIOMASS, ABUNDANCE, QBIOMASS and QABUNDANCE at
 the set level.

 path/data/length stores BIOMASS, ABUNDANCE, QBIOMASS and QABUNDANCE at the
 set level for LENGTH in 1 cm increments. 
 
 From\code{stratifyBiomass()}:data/stratified/folder/. The name of "folder" depends on the arguments
 \code{lengthbased} and \code{qadjusted}:

 if(lengthBased & qadjusted), then folder is named "lengthbased_qadj"

 if(lengthBased & !qadjusted), then folder is named "lengthbased_notqadj"

 if(!lengthBased & qadjusted), then folder is named "notlengthbased_qadj"

 if(!lengthBased & qadjusted), then folder is named "notlengthbased_notqadj"

 Inside "folder" is a folder for each entry in \code{area}. Within each area
 folder is a .RData file for each year from s.year to e.year called
 year.RData (object name \code{out}).

 \code{out} has 5 columns if \code{lengthbased = TRUE}: \code{SPECIES} ,
 \code{ID}, \code{FLEN}, \code{BIOMASS}, and \code{ABUNDANCE}.

 \code{out} has 4 columns if \code{lengthbased = FALSE}: \code{SPECIES},
 \code{ID}, \code{BIOMASS}, and \code{ABUNDANCE}.
}
\description{
Runs\code{biomassData()} to extract survey data and and run
 \code{stratifyBiomass()} to stratify biomass and abundance estimates.
}
\references{
Original code by DD.
}
