% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herringAtLength.R
\name{herringAtLength}
\alias{herringAtLength}
\title{Allocates 4X length-based herring index into sets and length classes
 based on the RV survey.}
\usage{
herringAtLength(path, s.year, e.year)
}
\arguments{
\item{path}{Filepath indicating where to create folders to store the extracted
data.}

\item{s.year}{Year for which to begin data extraction.}

\item{e.year}{Year for which to end data extraction.}
}
\description{
This function is called by \code{biomassData} to allocate the 4X
 herring index from XXXX into sets and length classes based on the proportion
 from the resaerch vessel survey.
}
\details{
User must define \code{channel = odbcConnect("ptran", uid = ###, pwd
 = ###)} in the global environment. This channel must have access to the XXXX
 databases.

 This function requires file path/extra info/stratweights/csv.
 stratweights.csv has two columns: \code{STRAT} and the corresponding
 \code{AREA}. Each \code{AREA} is converted to tow units using the conversion
 \code{TOW UNITS = AREA/((35./6080.2)*1.75)} until 1981 and \code{TOW UNITS =
 AREA/((41./6080.2)*1.75)} after 1981

 The 4X length-based herring index is based on XXXX. This data must be stored
 in path/extra info/**name.csv.

 The /code{ABUNDANCE} in **name.csv is allocated across the research vessel
 summer strata and length classes observed in the survey based on a
 stratification/destratification scheme:

 1. Stratify research vessel survey \code{ABUNDANCE} and determine the total
 \code{ABUNDANCE} over the whole area.

 2. Calculate proportion of biomass and abundance at each length (1 cm
 intervals).

 3. Calculate the proportion of the total \code{ABUNDANCE} that is
 contributed by each strata.

 4.  Calculate the proportion of the mean \code{ABUNDANCE} in each strata
 that is contributed by each set.

 5. Use the proportions in 2, 3, and 4 to destratify the 4X length-based
 herring index \code{ABUNDACE}.

 6. Use the mean weight of fish to calculate the \code{BIOMASS} in each set:
 \deqn{\code{BIOMASS} = \code{ABUNDANCE} * Weight_{Avg}}.

 (Notes:

 1. AC's code did not account for the different tow units after the gear
 change when de-stratifying the index, but DID when re-stratifying the
 biomass. The change in tow units is now accounted for in both procedures.

 2. AC's code did not account for the zero sets when calculating the mean
 abundance in each strata. This is now accounted for.)
}
\references{
Modified code from AC's ExtractIndicators/R/VPAherring.R
}
