% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invBiomass.R
\name{inv.biomass}
\alias{inv.biomass}
\title{Pretty sure this extracts the invertebrate biomass. But doesn't look
 like it is called by any other function so might not need it!
 @description Not sure why the default is \code{e.year = s.year}}
\usage{
inv.biomass(species, region, season, s.year, e.year = s.year)
}
\arguments{
\item{species}{Probably the species for which to extract data}

\item{region}{Probably the region for which to extract data}

\item{season}{Probably the season for which to extract data}

\item{s.year}{Year for which to begin data extraction.}

\item{e.year}{Year for which to end data extraction.}
}
\description{
Pretty sure this extracts the invertebrate biomass. But doesn't look
 like it is called by any other function so might not need it!
 @description Not sure why the default is \code{e.year = s.year}
}
\references{
Modified code from AC's ExtractIndicators/R/qBiomass.R
}
