% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landings_groupings.R
\docType{data}
\name{landings_groupings}
\alias{landings_groupings}
\title{Links NAFO sub-unit to area \code{ID}}
\format{A dataframe with 612 rows and 9 variables: \describe{
 \item{NAFOSUB}{} \item{NAFOSUBID}{} \item{NAFOID}{}
 \item{NAFO}{} \item{ESSWSSID}{} \item{ESSWSS}{} \item{SHELFID}{}
 \item{SHELF}{} \item{SHELF}{} }}
\source{
These landings groupings are from AC's file ExtractIndicators/extra
 info/landingsgroupings.csv.
}
\usage{
landings_groupings
}
\description{
Links NAFO sub-unit to area \code{ID} for the spatial scales of interest
(Shelf, ESS/WSS, NAFO divisions). Applied in \code{LANDdataframe()}.
}
\keyword{datasets}
