% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadfun.R
\name{loadfun}
\alias{loadfun}
\title{Sources all functions in the path.}
\usage{
loadfun(path)
}
\arguments{
\item{path}{The file path to the package}
}
\description{
I don't think I will need this when the package is built. This
 function sources all of the functions in SSdata/R, making them available for
 use by the package.
}
\references{
Modeified code from AC's ExtractIndicators/R/amc helpers.R
}
