% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LWdataframe.R
\name{LWdataframe}
\alias{LWdataframe}
\title{Formats length-weight data for use in the \code{marindicators} package}
\usage{
LWdataframe(path, s.year, e.year, areas = c("shelf", "esswss", "nafo",
  "strat"), update_LW = FALSE, csv = TRUE, rdata = TRUE)
}
\arguments{
\item{path}{The filepath to the data folder created by \code{extractLW()}.}

\item{s.year}{Year for which to begin data extraction.}

\item{e.year}{Year for which to end data extraction.}

\item{areas}{Areas (spatial scales) for which to format data. A separate
dataframe will be exported for each area. Options are "shelf", "esswss",
"nafo", "strat", or any combination of the four. Default is \code{areas =
c("shelf", "esswss", "nafo", "strat")}.}

\item{update_LW}{Logical parameter indicating whether to run
\code{extractLW()}. This can be time consuming, so if LW data is already
extracted, use the default \code{update_LW = FALSE}. If \code{update_LW =
TRUE}, user must define \code{channel} in the global environment (see
\code{Details}).}

\item{csv}{Logical value indicating whether to export dataframe as a .csv
file. Default is \code{csv = TRUE}.}

\item{rdata}{Logical value indicating whether to export dataframe as a .RData
file. Default is \code{rdata = TRUE}.}
}
\value{
This function creates a directory path/output/LengthWeight and stores
 the length-weight dataframe for each area in area_LengthWeight.RData (object
 name is \code{lw}) and/or area_LengthWeight.csv. The dataframe has 5
 columns: \code{ID}, \code{YEAR}, \code{SPECIES}, \code{LENGTH} (cm), and
 \code{WEIGHT} (g). These files are formatted for the \code{marindicators}
 package.
}
\description{
This function imports data from path/data/lenwgt adds column
 \code{YEAR}, and attaches labels for the spatial scales of interest (shelf,
 ESS/WSS, NAFO divisions, and/or strata) in column \code{ID}.
}
\details{
If \code{update_LW = TRUE}, user must define \code{channel =
 odbcConnect("ptran", uid = ###, pwd = ###)} in the global environment. This
 channel must have access to the gsinf and gsdet tables from the groundfish database.
}
\references{
Original code by DD.
}
