% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_LW.R
\name{plot_LW}
\alias{plot_LW}
\title{Plot length-weight data}
\usage{
plot_LW(species, year)
}
\arguments{
\item{species}{Species code for which to extract data. Must be one of the 88
species in the \code{catch_coefs} table in \code{biomassData}.}

\item{year}{Year for which to extract data. Must be one of the 88 species in
the \code{catch_coefs} table in \code{biomassData}.}
}
\value{
A plot of weight vs. length.
}
\description{
Plots length-weight data for given species and year. Displays the
 estimated values of a and b.
}
\references{
Modeified from AC's code in script qBiomass.
}
