% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qBiomassPostStrat.R
\name{qBiomassPostStrat}
\alias{qBiomassPostStrat}
\title{Applies length-based q-adjusted to biomass and abundance data (after
 stratification)}
\usage{
qBiomassPostStrat(aa, path)
}
\arguments{
\item{aa}{Stratified fishery independent survey data, as suplied from
\code{stratifyBiomass()}.}

\item{path}{Path indicating where to store stratified data.}
}
\value{
Returns the q-ajusted biomass and abundance to \code{bioamssData()}.
}
\description{
Notes from AC: 

biomass q adju using medians from harley and myers tech report

need to ensure 4X herring is not double q corrected
}
\references{
Modified code from AC's ExtractIndicators/R/qBiomassPostStrat.R
}
