% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVdataframe.R
\name{RVdataframe}
\alias{RVdataframe}
\title{Formats fishery independey survey data for use in the
 \code{marindicators} package}
\usage{
RVdataframe(path, s.year, e.year, areas = c("shelf", "esswss", "nafo",
  "strat"), lengthBased, csv = TRUE, rdata = TRUE)
}
\arguments{
\item{path}{The filepath to the data folder created by
\code{extractBiomass()}.}

\item{s.year}{Year for which to begin data extraction.}

\item{e.year}{Year for which to end data extraction.}

\item{areas}{Areas of interest. A separate dataframe will be exported for each
area. Default is \code{areas = c("shelf", "esswss", "nafo", "strat")}.}

\item{lengthBased}{Logical parameter indicating whether to format length-based
\code{lengthBased = TRUE} or not length-based data  \code{lengthBased =
FALSE}.}

\item{csv}{Logical value indicating whether to export dataframe as a .csv
file. Default is \code{csv = TRUE}.}

\item{rdata}{Logical value indicating whether to export dataframe as a .RData
file. Default is \code{rdata = TRUE}.}
}
\value{
This function creates a directory to save the exported data.

 If \code{lengthBased} is \code{TRUE}: path/output/RV/lengthBased and stores
 the dataframe in area_lengthBased.RData (object name is \code{RVdata})
 and/or area_RVdata.csv. The dataframe has 6 columns: \code{ID} (the area
 ID), \code{YEAR}, \code{SPECIES} (the research vessel species code),
 \code{LENGTH}, \code{BIOMASS}, and \code{ABUNDANCE}.

 OR

 If \code{lengthBased} is \code{FALSE}: path/output/RV/notlengthBased and
 stores the dataframe in area_notlengthBased.RData (object name is
 \code{RVdata}) and/or area_RVdata.csv. The dataframe has 5 columns:
 \code{ID} (the area ID), \code{YEAR}, \code{SPECIES} (the research vessel
 species code), \code{BIOMASS}, and \code{ABUNDANCE}.
}
\description{
This function imports data for each area from
 path/data/stratified/lengthbased or path/data/stratified/notlengthbased,
 adds column \code{YEAR}, and binds all years together.
}
\references{
Original code by DD.
}
