% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifyBiomass.R
\name{stratifyBiomass}
\alias{stratifyBiomass}
\title{Stratifies biomass and abundance estimates and then applies
 catchability correction factor (q)}
\usage{
stratifyBiomass(path, s.year, e.year, lengthBased, areas = c("shelf",
  "esswss", "nafo", "strat"))
}
\arguments{
\item{path}{Filepath indicating where to create folder to store the stratified
data.}

\item{s.year}{Year for which to begin data extraction.}

\item{e.year}{Year for which to end data extraction.}

\item{lengthBased}{Logical parameter indicating whether to return stratified
length-based \code{lengthBased = TRUE} or not length-based data
\code{lengthBased = FALSE}}

\item{areas}{Areas for which to stratify the fishery independent survey data.
A separate dataframe will be exported for each area. Default is \code{areas
= c("shelf", "esswss", "nafo", "strat")}.}
}
\value{
This function creates directories to store stratified data.

 path/data/stratified/lengthbased has a folder for each area. Within each
 folder is a .RData file for each year from s.year to e.year called
 year.RData (object name \code{out}). \code{out} has 5 columns:
 \code{SPECIES} , \code{FLEN}, \code{BIOMASS}, and \code{ABUNDANCE}.

 path/data/stratified/notlengthbased has a folder for each area. Within each
 folder is a .RData file for each year from s.year to e.year called
 year.RData (object name \code{out}). \code{out} has 4 columns:
 \code{SPECIES}, \code{BIOMASS}, and \code{ABUNDANCE}.
}
\description{
clean up this code
}
\references{
Modified code from AC's ExtractIndicators/R/stratifyBiomass.R
}
