% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vesselCorr.R
\name{vesselCorr}
\alias{vesselCorr}
\title{Applies vessel correction factor to key species}
\usage{
vesselCorr(x)
}
\arguments{
\item{x}{Fishery indepdendent survey data as extracted by \code{biomassData()}.}
}
\value{
This function returns the corrected biomass and abundance data to
 function \code{biomassData()}.
}
\description{
Called by \code{biomassData()} to apply vessel correction factor
 to key species.
}
\references{
Modified code from AC's ExtractIndicators/R/vesselCorrection.R.
}
\seealso{
Other RV functions: \code{\link{RVdataframe}},
  \code{\link{biomassData}}, \code{\link{extractRV}},
  \code{\link{qBiomass}}, \code{\link{stratifyBiomass}}
}
\concept{RV functions}
