% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassData.R
\name{biomassData}
\alias{biomassData}
\title{Extracts fishery independent data}
\usage{
biomassData(path, s.strat = 440, e.strat = 495, s.year, e.year,
  vessel.correction = TRUE)
}
\arguments{
\item{path}{Filepath indicating where to create folders to store the extracted
data.}

\item{s.strat}{Stratum for which to begin data extraction. Default is
\code{s.strat = 440}.}

\item{e.strat}{Stratum for which to end data extraction. Default is
\code{e.strat = 495}.}

\item{s.year}{Year for which to begin data extraction.}

\item{e.year}{Year for which to end data extraction.}

\item{vessel.correction}{Logical value indicating whether to apply vessel
correction to BIOMASS and ABUNDANCE. Default is vessel.correction = TRUE.}
}
\value{
This function creates directories to store extracted data. Within each
 folder, data for each year is saved in a separate RData file.

 path/data/aggregate stores BIOMASS, ABUNDANCE, QBIOMASS and QABUNDANCE at
 the set level.

 path/data/length stores BIOMASS, ABUNDANCE, QBIOMASS and QABUNDANCE at the
 set level for LENGTH in 1 cm increments.
}
\description{
Extracts annual length-based and not length-based
 (i.e., aggregated over length) data for the Scotian Shelf. QABUNDANCE and
 QBIOMASS are q-adjusted at the set level (i.e., before stratification).

 Herring biomass/abundance is treated separately from the other species.
}
\references{
Modified code from AC's ExtractIndicators/R/biomassData.R
}
