% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassData.R
\name{biomassData}
\alias{biomassData}
\title{Extracts fishery independent data}
\usage{
biomassData(path, s.strat = 440, e.strat = 495, s.year, e.year,
  vessel.correction = TRUE)
}
\arguments{
\item{path}{Filepath indicating where to create folders to store the extracted
data.}

\item{s.strat}{Stratum for which to begin data extraction. Default is
\code{s.strat = 440}.}

\item{e.strat}{Stratum for which to end data extraction. Default is
\code{e.strat = 495}.}

\item{s.year}{Year for which to begin data extraction.}

\item{e.year}{Year for which to end data extraction.}

\item{vessel.correction}{Logical value indicating whether to apply vessel
correction to calculation of BIOMASS and ABUNDANCE. Default is
vessel.correction = TRUE.}
}
\value{
This function creates directories to store extracted data. Within each
 folder, a separate RData file stores the data for each year from s.year to
 e.year, for strata from s.strat to e.strat.

 path/data/aggregate stores BIOMASS, ABUNDANCE, QBIOMASS and QABUNDANCE at
 the set level.

 path/data/length stores BIOMASS, ABUNDANCE, QBIOMASS and QABUNDANCE at the
 set level for lengths in 1 cm increments. Used in. . .

 path/data/lenwgt stores fish length at weight.
}
\description{
This function extracts annual length-based, aggregated, and
 length-weight data for the Scotian Shelf. Code can be cleaned up

 Herring biomass/abundance is treated separately from the other species.
}
\references{
Modified code from AC's ExtractIndicators/R/biomassData.R
}
