% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassData.R
\name{biomassData}
\alias{biomassData}
\title{Extracts and exports fishery independent biomass and abundance data}
\usage{
biomassData(path, s.strat = 440, e.strat = 495, s.year, e.year,
  vessel.correction = TRUE)
}
\arguments{
\item{path}{Filepath indicating where to create folders to store the extracted
data.}

\item{s.strat}{Stratum for which to begin data extraction. Default is
\code{s.strat = 440}. Code will stop with an error message if \code{s.strat
< 440} or \code{s.strat > 495}.}

\item{e.strat}{Stratum for which to end data extraction. Default is
\code{e.strat = 495}.  Code will stop with an error message if \code{e.strat
< 440} or \code{e.strat > 495}.}

\item{s.year}{Year for which to begin data extraction.}

\item{e.year}{Year for which to end data extraction.}

\item{vessel.correction}{Logical value indicating whether to apply vessel
correction to \code{BIOMASS} and \code{ABUNDANCE}. Default is
\code{vessel.correction} = TRUE.}
}
\value{
Creates directories to store extracted data.

 Not length-based data are stored in path/data/aggregate/. This folder
 includes an RData file for each year called year.RData (object name
 \code{dat}). \code{dat} has 9 columns: \code{MISSION}, \code{SETNO},
 \code{SPECIES}, \code{YEAR}, \code{STRAT}, \code{BIOMASS}, \code{ABUNDANCE},
 \code{QBIOMASS} and \code{QABUNDANCE}.

 Length-based data is stored in path/data/length/. This folder includes an
 RData file for each year called num_biom_at_length_year.RData (object name
 \code{out}). \code{out} has 10 columns: \code{MISSION}, \code{SETNO},
 \code{SPECIES}, \code{YEAR}, \code{STRAT}, \code{BIOMASS}, \code{ABUNDANCE},
 \code{QBIOMASS} and \code{QABUNDANCE}, and \code{FLEN}, where \code{FLEN} is
 length in 1 cm increments.
}
\description{
Extracts and exports annual q-adjusted and not q-adjusted,
 length-based (separated into 1 cm length classes) and not length-based
 (aggregated over length) biomass and abundance data from the Scotian Shelf
 summer research vessel surveys. q-adjustments are applied at the set level
 (before stratification).
}
\details{
User must define \code{channel = odbcConnect("ptran", uid = ###, pwd
 = ###)} in the global environment. This channel must have access to the
 gscat, gsdet, gsinf, and gs_lengths tables from the groundfish database and
 the nafo_strat table from the mfd_stomach database.

 Units: biomass - kg; abundance - numbers; length - cm.

 If \code{BIOMASS} or \code{ABUNDANCE} data is missing when the other is
 present, the missing field is filled in based on the relationship
 \eqn{Weight_{Avg} = \code{BIOMASS}/\code{ABUNDANCE}}. \eqn{Weight_{Avg}} is
 from a sql call to "mean_wts_for_fill_in." If \eqn{Weight_{Avg}} is not
 available for a species, it is assumed that \eqn{Weight_{Avg} = 0.5}.

 Abnormally high dogfish biomass estimates (> 8000 kg at the set level) are
 replaced with \code{BIOMASS = 0} kg.

 Abundance estimates of \code{Inf} in 2012 are replaced with \code{ABUNDANCE
 = 1}.

 Code will stop if any \code{BIOMASS} or \code{ABUNDANCE} estimates are
 \code{NA}.

 Estimates of zero are replaced with small values: \code{BIOMASS = 0} is
 replaced with \code{BIOMASS = 0.01}; \code{ABUNDANCE = 0} is replaced with
 \code{ABUNDANCE = 1}.
}
\references{
Modified code from AC's ExtractIndicators/R/biomassData.R
}
\seealso{
Other RV functions: \code{\link{RVdataframe}},
  \code{\link{extractRV}}, \code{\link{qBiomass}},
  \code{\link{stratifyBiomass}}, \code{\link{vesselCorr}}
}
\concept{RV functions}
