% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileDataframes.R
\name{compileDataframes}
\alias{compileDataframes}
\title{Exports fishery independent and commercial landings data}
\usage{
compileDataframes(path, s.year, e.year, areas.RV = c("strat", "nafo",
  "shelf", "esswss"), areas.land, csv = FALSE, rdata = TRUE)
}
\arguments{
\item{path}{Filepath indicating where to create folders to store the output.}

\item{s.year}{Year for which to begin data compilation.}

\item{e.year}{Year for which to end data compilation.}

\item{areas.RV}{Areas for which to compile fishery independent data. Default
is \code{areas.RV = c("strat", "nafo", "esswss", "shelf")}.}

\item{areas.land}{Areas for which to compile commercial landings data. Default
is \code{areas.RV = c("nafo", "esswss", "shelf")}.}

\item{csv}{Logical value indicating whether to export dataframe as a .csv
file. Default is \code{csv = TRUE}.}

\item{rdata}{Logical value indicating whether to export dataframe as a .RData
file. Default is \code{rdata = TRUE}.}
}
\value{
The output is saved in directories created by the functions:

 From \code{RVdataframe}: path/output/RV/area/lengthbased: annual
 length-based biomass and abundance data, q-corrected and then stratified,
 with columns \code{ID}, \code{YEAR}, \code{SPECIES}, \code{LENGTH},
 \code{BIOMASS} (UNITS) and \code{ABUNDANCE} (UNITS). File name is
 area_lengthBased.RData; object name is \code{RVdata}.

 path/output/RV/area/notlengthBased: annual biomass and abundance data,
 q-corrected and then stratified, with columns \code{ID}, \code{YEAR},
 \code{SPECIES}, \code{BIOMASS} (UNITS) and \code{ABUNDANCE} (UNITS). File
 name is area_notlengthBased.RData; object name is \code{RVdata}.

 From \code{LWdataframe()}: path/output/LengthWeight: An Rdata file of annual
 length-weight data with columns \code{ID}, \code{YEAR}, \code{SPECIES},
 \code{LENGTH} (cm), \code{WEIGHT} (UNITS). File name is
 area_LengthWeight.RData; object name is \code{lw}.

 From \code{LANDdatafrmae()}: path/output/Landings: An Rdata file of annual
 landings data with columns \code{ID}, \code{YEAR}, \code{SPECIES},
 \code{CATCH} (UNITS). The species codes in \code{SPECIES} are the RV codes.
 See \code{?LANDdataframe} for more information. File name is
 area_landings.RData; object name is land.
}
\description{
Exports fishery independent and commercial landings dataframes in
 a format suitable for the \code{marindicators} package. These dataframes
 correspond to \code{X}, \code{X_length}, \code{LenWt.table}, and \code{land}
 in the \code{marindicators} package.
}
\details{
Calls \code{biomassData()}, \code{stratifyBiomass()},
 \code{RVdataframe()}, \code{LWdataframe()}, and \code{LANDdataframe()}.

 User must define \code{channel = odbcConnect("ptran", uid = ###, pwd = ###)}
 in the global environment. This channel must have access to the XXXX (RV)
 and NAFO, ZIF, and MARFIS (commercial) databases.
}
\references{
Original code by DD.
}
