% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileDataframes.R
\name{compileDataframes}
\alias{compileDataframes}
\title{Exports fishery independent and commercial landings data}
\usage{
compileDataframes(path, s.year, e.year, areas.RV = c("strat", "nafo",
  "shelf", "esswss"), areas.land, csv = F, Rdata = TRUE)
}
\arguments{
\item{path}{Filepath indicating where to create folders to store the output.}

\item{s.year}{Year for which to begin data compilation.}

\item{areas.RV}{Areas for which to compile fishery independent data. Default
is \code{areas.RV = c("strat", "nafo", "shelf", "esswss")}.}

\item{s.year}{Year for which to end data compilation.}

\item{areas.RV}{Areas for which to compile commercial landings data. Default
is \code{areas.RV = c("nafo", "shelf", "esswss")}.}
}
\value{
The output is saved in directories created by the function:

 path/output/Landings: An Rdata file of annual landings data with columns
 \code{ID}, \code{YEAR}, \code{SPECIES}, \code{CATCH}. The species codes in
 \code{SPECIES} are the RV codes. See ?LANDdataframe for more information.
 File name is area_landings.RData; object name is land.

 path/output/LengthWeight: An Rdata file of annual length-weight data with
 columns \code{ID}, \code{YEAR}, \code{SPECIES}, \code{LENGTH},
 \code{WEIGHT}. File name is area_LengthWeight.RData; object name is
 \code{lw}.

 path/output/RV/lengthBased: annual length-based biomass and abundance data,
 stratified and then q-corrected, with columns \code{ID}, \code{YEAR},
 \code{SPECIES}, \code{LENGTH}, \code{BIOMASS} and \code{ABUNDANCE}. File
 name is area_lengthBased.RData; object name is \code{RVdata}.

 path/output/RV/notlengthBased: annual biomass and abundance data, stratified
 and then q-corrected, with columns \code{ID}, \code{YEAR}, \code{SPECIES},
 \code{BIOMASS} and \code{ABUNDANCE}. File name is area_notlengthBased.RData;
 object name is \code{RVdata}.
}
\description{
Exports fishery independent and commercial landings
 dataframes in a format suitable for the \code{marindicators} package.
}
\references{
Original code by DD.
}
