% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineAreas.R
\name{defineAreas}
\alias{defineAreas}
\title{Assigns area ID based on strata}
\usage{
defineAreas(dat, area)
}
\arguments{
\item{dat}{A dataframe that includes the column \code{STRAT}, which has
entries corresponding to the strata on the Scotian Shelf (e.g., values from
440 to 495).}

\item{area}{Character string indicating the spatial scale for which to assign
areas IDs. Options are: \code{"strat"}, \code{"nafo"}, \code{"esswss"},
\code{"shelf"}, \code{"indiseas"}, \code{"pollockwest"},
\code{"pollockeast"}, \code{"unit2redfish"}, \code{"unit3redfish"},
\code{"silverhake"}.}
}
\value{
The function returns \code{dat} with an extra column: \code{ID}
}
\description{
Adds column \code{ID} to a dataframe that already has a column
 \code{STRAT}, and then assigns an area ID based on the strata number in
 \code{STRAT}.
}
\references{
Modified code from AC's ExtractIndicators/R/defineGroups.R.
}
