% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBiomass.R
\name{extractBiomass}
\alias{extractBiomass}
\title{Calls functions to extract and stratify fishery independent data}
\usage{
extractBiomass(path, s.year, e.year, areas)
}
\arguments{
\item{path}{Filepath indicating where to create folders to store the extracted
data.}

\item{s.year}{Year for which to begin data extraction.}

\item{e.year}{Year for which to end data extraction.}
}
\value{
This function creates directories to store extracted data. Within each
 folder, a separate RData file stores the data for each year from s.year to
 e.year, for strata from s.strat to e.strat.

 From \code{biomassData()}:
 path/data/aggregate stores BIOMASS, ABUNDANCE, QBIOMASS and QABUNDANCE at
 the set level. 
 
 path/data/length stores BIOMASS, ABUNDANCE, QBIOMASS and
 QABUNDANCE at the set level for lengths in 1 cm increments. Used in. . .
 
 path/data/lenwgt stores fish length at weight.
 
 From \code{stratifyBiomass()}:
 path/data/stratfied/lengthBased stores the length-based stratified,
 q-adjusted (post-stratification) BIOMASS and ABUNDANCE.
 
 path/data/stratfied/notlengthBased stores the not length-based stratified,
 q-adjusted (post-stratification) BIOMASS and ABUNDANCE.
}
\description{
Calls\code{biomassData()} and \code{stratifyBiomass()}. For more
 informations, see the help files for those functions.
}
\references{
Original code by DD.
}
