% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLAND.R
\name{extractLAND}
\alias{extractLAND}
\title{Extracts commercial landings data from historic and current databases}
\usage{
extractLAND(path, e.year)
}
\arguments{
\item{path}{Filepath indicating where to create the folder to store the
extracted data.}

\item{e.year}{Year for which to end data extraction.}
}
\value{
This function creates directory path/data/landings and stores file
 landings.RData (object name is \code{landings}). \code{landings} has 14
 columns: \code{SPECIES}, \code{ALLNAMES}, \code{YEAR}, \code{NAFO_UNIT},
 \code{CATCH} (tonnes), and 9 landings groups. A value of \code{1} in a
 landings group column indicates that the corresponding \code{SPECIES}
 belongs to that group. Species may belong to more than one group. Species
 codes are the commercial landings codes.
}
\description{
Extracts commercial landings data from NAFO (1968 - 1985), ZIF
 (1986 - 2002) and MARFIS (2003 - present) databases.
}
\details{
User must define \code{channel = odbcConnect("ptran", uid = ###, pwd
 = ###)} in the global environment. This channel must have access to: the
 nafo_summary and nafo_area_codes tables from the COMLAND database, the
 sub_trips_XXXX and identified_catches tables from the cl database, the
 marfis_catch_effort table from mfd_obfmi database, and the
 indiseas_marfis2allcodes table from the gomezc database.

 Units: tonnes 

 This function defines and then calls three functions: \code{NAFO()},
 \code{ZIF()}, and \code{MARFIS()}.

 Extracts data from 1968 to \code{e.year}.

 Species are assigned to commercial landings groups based on
 \code{commercial_groups}. Type \code{?commercial_groups} for more
 information.
}
\references{
Modified code from AC's ExtractIndicators/R/getLandings.R
}
\seealso{
Other LAND functions: \code{\link{LANDdataframe}}
}
\concept{LAND functions}
