% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLW.R
\name{extractLW}
\alias{extractLW}
\title{Extracts annual length-weight data}
\usage{
extractLW(path, s.year, e.year)
}
\arguments{
\item{path}{Filepath indicating where to create folders to store the extracted
data.}

\item{s.year}{Year for which to begin data extraction.}

\item{e.year}{Year for which to end data extraction.}
}
\value{
This function creates a directory to store extracted data:
 path/data/lenwgt stores a separate file for each year, called lwyear.RData
 (object name \code{wt}). \code{wt} has 4 columns: \code{STRAT},
 \code{SPECIES}, \code{FLEN} (length in cm), \code{FWT} (weight in UNITS).
}
\description{
Extracts annual length-weight data from fishery independent
 surveys on the Scotian Shelf.
}
\details{
User must define \code{channel = odbcConnect("ptran", uid = ###, pwd
 = ###)} in the global environment. This channel must have access to the XXXX
 database.
}
\references{
Modified code from AC's ExtractIndicators/R/biomassData.R
}
