% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herringAggregate.R
\name{herringAggregate}
\alias{herringAggregate}
\title{Not really sure what this does}
\usage{
herringAggregate(path1)
}
\arguments{
\item{path1}{Filepath indicating where extra info folder is stored}
}
\description{
Documenting the function so I can build the package.
}
\references{
Modified code from AC's ExtractIndicators/R/VPAherring.R
}
