% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herringAtLength.R
\name{herringAtLength}
\alias{herringAtLength}
\title{Not really sure what this does}
\usage{
herringAtLength(path, s.year, e.year)
}
\arguments{
\item{path}{Filepath indicating where to create folders to store the extracted
data.}

\item{s.year}{Year for which to begin data extraction.}

\item{e.year}{Year for which to end data extraction.}
}
\description{
Documenting the function so I can build the package.
}
\references{
Modified code from AC's ExtractIndicators/R/VPAherring.R
}
