% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDdataframe.R
\name{LANDdataframe}
\alias{LANDdataframe}
\title{Formats commercial landings data for use in the \code{marindicators}
 package}
\usage{
LANDdataframe(path, areas = c("shelf", "esswss", "nafo"), update_LAND,
  e.year = NULL, csv = TRUE, rdata = TRUE)
}
\arguments{
\item{path}{The filepath to the data folder created by
\code{extractLandings()}.}

\item{areas}{Areas of interest. A separate landings dataframe will be exported
for each area. Default is \code{areas = c("shelf", "esswss", "nafo")}}

\item{update_LAND}{Logical parameter indicating whether to run
\code{extractLAND()}. This may be time consuming, so if data is already
extracted, use the default \code{update_LAND = FALSE}. If \code{update_LAND
= TRUE}, user must define \code{channel} in the global environment (see
\code{Details}).}

\item{e.year}{If \code{update_LAND = TRUE}, \code{e.year} is the final year
for which to extract the landings data. Default is \code{e.year = NULL}.}

\item{csv}{Logical value indicating whether to export landings dataframe as a
.csv file. Default is \code{csv = TRUE}.}

\item{rdata}{Logical value indicating whether to export landings dataframe as
a .RData file. Default is \code{rdata = TRUE}.}
}
\value{
This function creates a directory output/Landings and stores the
 landings dataframe in area_land.RData (object name is \code{land}) and/or
 area_land.csv. The dataframe has 4 columns: \code{ID} (the area ID),
 \code{YEAR}, \code{SPECIES} (the research vessel species code) and
 \code{CATCH} (UNITS).
}
\description{
This function imports data from
 path/data/landings.landings.RData, attaches labels for the spatial scales of
 interest ("shelf", "esswss", and "nafo") in column \code{ID}, and replaces
 commercial species codes with the research vessel species codes.
}
\details{
If \code{update_LAND = TRUE}, user must define \code{channel =
 odbcConnect("ptran", uid = ###, pwd = ###)} in the global environment. This
 channel must have access to the XXXX databases.

 Area ID's are assiged from the file path/extra info/landingsgroupings.csv.

 Commercial and research vessel species codes are matched from the file
 path/extra info/SpeciesCodes.csv. The function accounts for the proportion
 of landings of species with more than one RV code but one commercial code.
}
\references{
Original code by DD.
}
