% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LWdataframe.R
\name{LWdataframe}
\alias{LWdataframe}
\title{Formats length-weight data for use in the \code{marindicators} package}
\usage{
LWdataframe(path, s.year, e.year, areas = c("shelf", "esswss", "nafo",
  "strat"), csv = TRUE, rdata = TRUE)
}
\arguments{
\item{path}{The filepath to the data folder created by
\code{extractBiomass()}.}

\item{s.year}{Year for which to begin data extraction.}

\item{e.year}{Year for which to end data extraction.}

\item{areas}{Areas of interest. A separate dataframe will be exported for each
area. Default is \code{areas = c("shelf", "esswss", "nafo", "strat")}.}

\item{csv}{Logical value indicating whether to export dataframe as a .csv
file. Default is \code{csv = TRUE}.}

\item{rdata}{Logical value indicating whether to export dataframe as a .RData
file. Default is \code{rdata = TRUE}.}
}
\value{
This function creates a directory path/output/LengthWeight to save the
 exported data. A separate .RData file (area_LengthWeigth.RData, object name
 \code{lw}) and/or .csv file (area_LengthWeight.csv) is exported for each area.
}
\description{
This function imports data from path/data/lenwgt adds column
 \code{YEAR}, and attaches labels for the spatial scales of interest
 ("shelf", "esswss", "nafo", "strat") in column \code{ID}.
}
\references{
Original code by DD.
}
