% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qBiomass.R
\name{qBiomass}
\alias{qBiomass}
\title{Applies length-based q-adjusted to biomass and abundance data (before
 stratification)}
\usage{
qBiomass(species, year, area = "4VWX", fun_group = NA, q = 0,
  len_corr = 1, nums = F, a.b.plot = F, by.set = T,
  by.length = T)
}
\arguments{
\item{species}{Species from catchability coefficients table
(\code{catch_coef}) in \code{biomassData()}.}

\item{year}{Year to calculate the data}

\item{area}{Default is \code{area = "4VWX"}}

\item{fun_group}{Functional group from catchability coefficients table
(\code{catch_coef}) in \code{biomassData()}.}

\item{q}{Initial q-correction from catchability coefficients table
(\code{catch_coef}) in \code{biomassData()}.}

\item{len_corr}{Initial length correction from catchability coefficients table
(\code{catch_coef}) in \code{biomassData()}.}

\item{nums}{Not really sure what this does.}

\item{a.b.plot}{Option to make length-weight plot.}

\item{by.set}{Return data by set. (?) Default is \code{by.set = TRUE}.}

\item{by.length}{Return data by set. (?) Default is \code{by.length = TRUE}.}
}
\value{
Returns the q-ajusted biomass and abundance to ]code{bioamssData()}.
}
\description{
I'm not sure if I need this function since we are doing postStat.
 But it sets up the extraction of the biomass/abundance data in
 biomassData(). Note from AC: biomass q adju using medians from harley and
 myers tech report
}
\references{
Modified code from AC's ExtractIndicators/R/qBiomass (function
 \code{biomass_q_adj()})
}
