% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVdataframe.R
\name{RVdataframe}
\alias{RVdataframe}
\title{Formats fishery independey survey data for use in the
 \code{marindicators} package}
\usage{
RVdataframe(path, s.year, e.year, areas = c("shelf", "esswss", "nafo",
  "strat"), lengthbased, qadjusted, update_RV = FALSE, csv = TRUE,
  rdata = TRUE)
}
\arguments{
\item{path}{The filepath to the data folder created by
\code{extractBiomass()}.}

\item{s.year}{Year for which to begin data extraction.}

\item{e.year}{Year for which to end data extraction.}

\item{areas}{Areas of interest. A separate dataframe will be exported for each
area. Default is \code{areas = c("shelf", "esswss", "nafo", "strat")}.}

\item{lengthbased}{Logical parameter indicating whether to format length-based
\code{lengthbased = TRUE} or not length-based data \code{lengthbased =
FALSE}.}

\item{qadjusted}{Logical parameter indicating whether}

\item{update_RV}{Logical parameter indicating whether to run
\code{extractRV()}. Extracting the abundance and biomass data for all years
and areas is time consuming, so if data is already extracted, use the
default \code{update_biomass = FALSE}.}

\item{csv}{Logical value indicating whether to export dataframe as a .csv
file. Default is \code{csv = TRUE}.}

\item{rdata}{Logical value indicating whether to export dataframe as a .RData
file. Default is \code{rdata = TRUE}.}
}
\value{
This function creates a directory path/output/RV to save the exported
 data. In the RV folder is a folder for each entry in \code{area}. In each
 area folder is a csv and/or Rdata file for the specified combination of
 \code{lengthbased} and \code{qadjusted} (object name \code{RVdata}. These
 files are formatted for the \code{marindicators} package.

 If \code{lengthbased = TRUE}, \code{RVdata} has 6 columns: \code{YEAR},
 \code{SPECIES}, \code{ID}, \code{LENGTH}, \code{BIOMASS}, and \code{ABUNDANCE}.
 
 If \code{lengthbased = FALSE}, \code{RVdata} has 5 columns: \code{YEAR},
 \code{SPECIES}, \code{ID}, \code{BIOMASS}, and \code{ABUNDANCE}.
}
\description{
This function imports data for each area from
 path/data/stratified/lengthbased or path/data/stratified/notlengthbased,
 adds column \code{YEAR}, and binds all years together.
}
\references{
Original code by DD.
}
