% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVdataframe.R
\name{RVdataframe}
\alias{RVdataframe}
\title{Formats fishery independent survey data for use in the
 \code{marindicators} package}
\usage{
RVdataframe(path, s.year, e.year, areas = c("shelf", "esswss", "nafo",
  "strat"), lengthbased, qadjusted, update_RV = FALSE, csv = TRUE,
  rdata = TRUE)
}
\arguments{
\item{path}{The filepath to the /data folder.}

\item{s.year}{Year for which to begin data extraction.}

\item{e.year}{Year for which to end data extraction.}

\item{areas}{Areas (spatial scales) for which to format data. A separate
dataframe will be exported for each area. Options are "shelf", "esswss",
"nafo", "strat", or any combination of the four. Default is \code{areas =
c("shelf", "esswss", "nafo", "strat")}.}

\item{lengthbased}{Logical parameter indicating whether to return stratified
length-based data (\code{lengthBased = TRUE}) or not length-based data
(\code{lengthBased = FALSE}).}

\item{qadjusted}{Logical parameter indicating whether to return q-adjusted
biomass and abundance data (\code{qadjusted = TRUE}) or not q-adjusted
biomass and abundance data (\code{qadjusted = FALSE}). q-adjustments are
applied at the set level.}

\item{update_RV}{Logical parameter indicating whether to run
\code{extractRV()}. Extracting the abundance and biomass data for all years
and areas is time consuming, so if data is already extracted, use the
default \code{update_biomass = FALSE}. If \code{update_RV = TRUE}, user
must define \code{channel} in the global environment (see \code{Details}).}

\item{csv}{Logical value indicating whether to export dataframe as a .csv
file. Default is \code{csv = TRUE}.}

\item{rdata}{Logical value indicating whether to export dataframe as a .RData
file. Default is \code{rdata = TRUE}.}
}
\value{
This function creates a directory path/output/RV/area for each entry
 in \code{area}. In each area folder is a csv and/or RData file for the
 specified combination of \code{lengthbased} and \code{qadjusted} (object
 name \code{RVdata}). These files are formatted for the \code{marindicators}
 package.

 If \code{lengthbased = TRUE}, then \code{RVdata} has 6 columns: \code{YEAR},
 \code{SPECIES}, \code{ID}, \code{LENGTH} (cm), \code{BIOMASS} (kg), and
 \code{ABUNDANCE} (numbers).

 If \code{lengthbased = FALSE}, then \code{RVdata} has 5 columns:
 \code{YEAR}, \code{SPECIES}, \code{ID}, \code{BIOMASS} (kg), and
 \code{ABUNDANCE} (numbers).
}
\description{
This function imports data exported by \code{stratifyBiomass()},
 adds column \code{YEAR}, and binds all years together.
}
\details{
If \code{update_RV = TRUE}, user must define \code{channel =
 odbcConnect("ptran", uid = ###, pwd = ###)} in the global environment. This
 channel must have access to the gscat, gsdet, gsinf, and gs_lengths tables
 from the groundfish database and the nafo_strat table from the mfd_stomach
 database.
}
\references{
Original code by DD.
}
