% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifyBiomass.R
\name{stratifyBiomass}
\alias{stratifyBiomass}
\title{Stratifies biomass and abundance estimates}
\usage{
stratifyBiomass(path, s.year, e.year, lengthbased, qadjusted,
  areas = c("esswss", "nafo", "shelf", "strat"))
}
\arguments{
\item{path}{Filepath indicating where to create folder to store the stratified
data.}

\item{s.year}{Year for which to begin data extraction.}

\item{e.year}{Year for which to end data extraction.}

\item{lengthbased}{Logical parameter indicating whether to return stratified
length-based data (\code{lengthBased = TRUE}) or not length-based data
(\code{lengthBased = FALSE}).}

\item{qadjusted}{Logical parameter indicating whether to return q-adjusted
biomass and abundance data (\code{qadjusted = TRUE}) or not q-adjusted
biomass and abundance data (\code{qadjusted = FALSE}). q-adjustments are
applied at the set level.}

\item{areas}{Areas (spatial scales) for which to stratify the fishery
independent survey data. A separate dataframe will be exported for each
area. Options are "shelf", "esswss", "nafo", "strat", or any combination of
the four. Default is \code{areas = c("shelf", "esswss", "nafo", "strat")}.}
}
\value{
This function creates a directory to store the stratified data:
 data/stratified/folder/. The name of "folder" depends on the arguments
 \code{lengthbased} and \code{qadjusted}:

 \code{if(lengthBased & qadjusted)}, then folder is named "lengthbased_qadj"

 \code{if(lengthBased & !qadjusted)}, then folder is named
 "lengthbased_notqadj"

 \code{if(!lengthBased & qadjusted)}, then folder is named
 "notlengthbased_qadj"

 \code{if(!lengthBased & qadjusted)}, then folder is named
 "notlengthbased_notqadj"

 Inside "folder" is a folder for each entry in \code{area}. Within each area
 folder is a .RData file for each year from s.year to e.year called
 year.RData (object name \code{out}).

 \code{out} has 5 columns if \code{lengthbased = TRUE}: \code{SPECIES} ,
 \code{ID}, \code{FLEN}, \code{BIOMASS}, and \code{ABUNDANCE}.

 \code{out} has 4 columns if \code{lengthbased = FALSE}: \code{SPECIES},
 \code{ID}, \code{BIOMASS}, and \code{ABUNDANCE}.
}
\description{
Stratifies the biomass and abundance estimates based on the
 weights in \code{strat_weights} (type \code{?strat_weights} into the console
 for more information). User can choose whether to return length-based and/or
 q-corrected values.
}
\references{
Modified code from AC's ExtractIndicators/R/stratifyBiomass.R
}
