% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vesselCorr.R
\name{vesselCorr}
\alias{vesselCorr}
\title{Applies vessel correction factor to key species}
\usage{
vesselCorr(x)
}
\arguments{
\item{x}{Fishery indepdendet survey data as extracted in the At Lwngth section
of \code{biomassData()}.}
}
\value{
This function returns the corrected biomass and abundance data to
 function \code{biomassData()}.
}
\description{
Notes from AC: correct for different survey vessels (after, L.P
 Fanning 1985): to obtain Alfred Needler comparable units: Lady Hammond
 (1982) and Alfred Needler (1983 to present) used a Western IIA Otter Trawl
 whereas The A.T. Cameron used a Yankee 35 ft trawl between 1970 to 1981 for
 total numbers need to use appropriate wingspreads 35ft for yankee and 41ft
 for western (From Manning 1985; reinforced by SJ Smith 2013).
}
\references{
Original code by DD.
}
